run https on local to work with location API
I am keeping the LocationButton react Functional component for now.
It was created in the first place to see why Stimulus wasn't working
with the GeoLocation API. It turns out that it will only work via
HTTPS. After creating a local cert and key I was able to get Stimulus
to work with the Safari browser forcing the rails to start up the server
with SSL.

It may end up that Stimulus isn't the right direction for these
asynchronous calls, and I will switch them into react components.