Moved to puma-dev for local https
The original solution of creating a self signed certificate worked
intermittently on Safari. It worked all the time on Firefox, but the
mutationObserver was undefined which I believe was a wrapper written
in Stimulus. Without that method React components would not render, and
my Stimulus actions with vanilla HTML would not trigger. Puma-Dev solved
all these issues by creating certificates in real time that Chrome
accepts.