refactor to anonomys arrow callback function to bind this
Dealing with a function that we didn't define has been difficult.
One solution would be to use call or apply to send the this context into
the callback. Something in Stimulus prevents that from happening as they
handle the binding of this in the Stimulus actions.

So instead of declaring a function in it's own context I decided to give
arrow functions a try since they don't create a new execution context
from the object to the left of the dot, and carry the execution context
from where they are declared.