map over json to get hash for views
The array that came in was an array of json objects. I could not parse
the entire array so I mapped into the original array parsing the index
json strings. This required changing the render partial into passing a
local so that I could access individual indexes for rendering.

I also added the html for the views that are rendering.